Update patch set 9
Patch Set 9:

> the nexus stuff doesn't apply to opnfv, but I wont hurt. I think
 > the right time to integrate this stuff would be when we move to a
 > new jenkins server.. (as it installs jjb) but how do we keep our
 > releng repo in sync with this going forward?

Since global-jjb is in a different Gerrit system than all the project Gerrit's we cannot depend on submodule subscriptions to automatically move it along but that's probably a good thing.

I plan to treat global-jjb as a software project. It will have releases. So you check "checkout" new versions from git tags when you want to upgrade. We'll have maintenance and release tags.

Patch-set: 9