Update patch set 1
Patch Set 1: Code-Review-2

No. This only happens on merge jobs that have a lot of deletes. Delete is not a common occurrence, we should not set a long timeout for this edge case as the majority of time it should pass in less than 10 minutes.

The reason deletes take long is because it's deleting all job artifacts at the same time. But we do not regularly delete hundreds of jobs. If the edge case happens just rerun the job until all affected jobs are deleted.

Patch-set: 1
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-2