Update patch set 2
Patch Set 2: Code-Review+2

> (2 comments)
 > 
 > How could we test a change like this? Something like update the
 > gitmodules hash to include it, in a GH project that uses this JJB,
 > and propose a test change or push to a ~sandbox?

This is actually already an established pattern for us in this repo. Take a look at the lf-ci-jobs.yaml to see what I'm talking about. As for testing it, we don't have any test jobs in a GitHub project for validating global-jjb just yet. My read on this is that it's good though.

Patch-set: 2
Label: Code-Review=+2