Update patch set 21
Patch Set 21: Code-Review-1

(8 comments)

So if we look at
https://nexus.opendaylight.org/content/repositories/opendaylight-yum-epel-6-x86_64/

I see that rpms are pushed to the "rpm" directory and srpms are pushed to the "srpm" directory.

Assuming we want to maintain the same structure this means group_id should be "rpm" for rpms and "srpm" for srpms. Seems like the rpm Nexus plugin doesn't actually care how the rpm repo is structured though so we could also consider using group_id as "org.opendaylight.integration.packaging" format as well. Not sure what the best plan here is.

Andy any suggestions?

Patch-set: 21
Label: Code-Review=-1