Update patch set 21
Patch Set 21:

> (8 comments)
 > 
 > So if we look at
 > https://nexus.opendaylight.org/content/repositories/opendaylight-yum-epel-6-x86_64/
 > 
 > I see that rpms are pushed to the "rpm" directory and srpms are
 > pushed to the "srpm" directory.
 > 
 > Assuming we want to maintain the same structure this means group_id
 > should be "rpm" for rpms and "srpm" for srpms. Seems like the rpm
 > Nexus plugin doesn't actually care how the rpm repo is structured
 > though so we could also consider using group_id as
 > "org.opendaylight.integration.packaging" format as well. Not sure
 > what the best plan here is.
 > 
 > Andy any suggestions?

RPM repos doesn't really care about the directory structure. Nor do apt repos for that matter, with just a few exceptions on metadata for both of them. It's fairly common practice to split the RPMs and SRPMs into separate directory structures which is why we were doing so previously.

However, with the way the artifacts will go up as a directory structure per artifactId I don't think maintaining that separation is of major value as it's generally just to help lessen the load on the server if you're browsing the indexes directly as it would in a normal repo drop 50% of the files it would have list.

Finally, since RPMs and DEbs are a flat namespace for artifact names just having everything go up at to root level of the repo would make sense, the only drawback is our current targetting permissions in Nexus, so in this case the GID should still end up being the project GID that produced the artifact.

Patch-set: 21
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0