Update patch set 1
Patch Set 1: Code-Review+1

I investigated this yesterday and it has nothing to do with the its plugins. It's purely commentlink related in Gerrit.

The problem is that Gerrit already converts URLs into links and commentlink feature searches for patterns and turns those patterns into links. The issue is that since Gerrit's automatic link parsing creates:

 
 <a href="URL">https://someurl/YOUR_PATTERN</a>
 
When a commentlink pattern hits it additionally inserts a link inside of the <a href></a> and embedded linking is invalid in HTML. So we get something like:
 
 <a href="URL">https://someurl/<a href="commentlink_url">YOUR_PATTERN</a></a>
 
Which is obviously incorrect can causes html that doesn't look quite right. The correct fix would require Gerrit to not commentlink if it detects a link is already present. Or we can find a regex pattern that does not trigger if it detects that it's in between an HTML anchor. There's an open bug against Gerrit about this here:

https://bugs.chromium.org/p/gerrit/issues/detail?id=3719

Patch-set: 1
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=+1