Update patch set 2
Patch Set 2: Code-Review-2

(2 comments)

This patch ignores the problem stated in the trello card https://trello.com/c/uSwEfMmP/542-make-build-rpm-stream-console-output-less-verbose by hiding the logs in another log. The transfer details are still printed in the file it's being redirected to.

To disable the transfer information from Maven output we need to pass 2 parameters:

1) --batch-mode
2) -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn

There is a standard set of variables we typically pass to Maven calls as defined in releng/builder's defaults.yaml as well as in global-jjb shell script here:

https://github.com/lfit/releng-global-jjb/blob/master/shell/common-variables.sh

Patch-set: 2
Label: Code-Review=-2