Update patch set 2
Patch Set 2:

> Decided to go with 45 minutes for the merge jobs to keep our job
 > runs tighter. I modified this patch so not going to merge it
 > myself.

I guess 45 mins is not enough for ODL, since we are building devstack. I think we might need to bump up the timeouts to 60/90 mins to be on safer side.

02:14:32 Build timed out (after 60 minutes). Marking the build as failed.
https://jenkins.opendaylight.org/releng/view/packer/job/builder-packer-merge-centos-devstack-pre-pip-newton/1/consoleFull

Why is devstack timing out after 60 mins when its expected to timeout after 45 mins ? 

https://jenkins.opendaylight.org/releng/view/packer/job/builder-packer-merge-centos-java-builder/1/console

the Java builder image timeout for 45 mins correctly:
02:00:18 Build timed out (after 45 minutes). Marking the build as failed.

Patch-set: 2