Update patch set 2
Patch Set 2:

We can declare another config-file-provider which will copy another one over when it's needed again. It doesn't cost us anything and it ensures that the files only exist for as long as they are needed. I see global-jjb macros as one-shot standalone macros.

We do not want to trust downstream template designers to remember to clean up after themselves and the scripts are all pretty similar. Having a multi-step macro that someone has to remember to copy 2 of has a high chance of folks forgetting the copy the cleanup step because their stuff will work even if they don't cleanup.

I think what this patch is trying to accomplish can be done in another macro that does similar to this one:

1) cfp to pull settings
2) run it's scripts (however many there may be)
3) cleanup cfp

This kind of stuff is why macros exist. It makes it easy to make standalone oneshot kind of commands like this. Using it to call a single script run is kind of pointless because then we could just call a one liner like "- shell: !include-raw: script.sh" instead of the 5+ lines it takes to declare a macro and use it.

Patch-set: 2