Update patch set 2
Patch Set 2: Code-Review-2

It's done purposely this way because I don't see a reason for a project non-lftools devs to want to use gerrit patches or git branches. They should always be installing versions of lftools.

We could patch this script to allow end user to actually pick released versions of lftools though. Right now it's hardcoded to use <1.0.0. I'll probably want to refactor LFTOOLS_VERSION variable in this case so that it would represent a version number. Maybe a different variable name to represent the dev builds.

Let me think about this a bit more.

Patch-set: 2
Label: Code-Review=-2