Convert lftools install to a macro
This allows us to not require LFTOOLS_VERSION defined in parameters
section so keeps that aspect of the code cleaner. I expect the only
folks who might be interested in setting LFTOOLS_VERSION are admins and
lftools developers. The version can still be overrided at the project
level by passing 'lftools-version: 1.0.0' in their project definition.

Shell scripts that want to use lftools after calling the macro will
need to also `source /tmp/v/lftools/bin/activate' to pull in the
virtualenv for lftools.

Change-Id: I74023397a4d808bd149b5952a73560562afb42a4
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>