Create job to lock|unlock branches via Gerrit
A contributor can leave a comment "lock|unlock branch" in Gerrit comment
which triggers a job to submit a patch to Gerrit for the project to lock
or unlock their branch.

Locking refers to removing the permission to click the Submit button in
the project's branch effectively disallowing new patches to be merged
into the repo. This job simply generates and submits a patch. A project
committer still needs to approve it.

Issue: RELENG-424
Change-Id: Id489d931037c76bdcad032ec44b0ce748bdffe81
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>