Update patch set 4
Patch Set 4:

If we are allowing locking project branches a self-service process, there is a chance that the branch would not be locked in time for the RC cut off date. Do we just manually lock all-project's if that happens?

Patch-set: 4
Reviewer: Gerrit User 29 <29@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Verified=0