Allow project-pattern to be overridable
While the ** pattern is good for global docs projects like the
OpenDaylight docs project. If projects have their own standalone
documentation such as lftools we need to be able to set project-pattern
to the exact project. This ensures that Gerrit won't start a lftools
build when a trigger happens on any other project in the Gerrit system.

Change-Id: I46c7d0d45eb7ea82230437648ff23e81baab07f1
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>