Add job to deploy to sandbox
This job allows a contributor to leave a comment for example
'jjb-deploy builder-jjb-*' to deploy a job (or several jobs if wildcards
are used) to the Jenkins Sandbox.

This allows jjb contributors to no longer have to install jjb on their
local systems in order to push the job to the Jenkins Sandbox.

ATTN: GitHub version of the job is untested and likely does NOT work as
      currently designed.

Issue: RELENG-373
Change-Id: I4f26c640804a0b7db1a0087cbd805b9263821326
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>