Add license check job using zxiiro/lhc
Create a job using https://github.com/zxiiro/license-header-checker as
the license checker tool for projects.

Start with only a Gerrit job for now as the GitHub job will need to be
designed in a different way as there's no way to make a single job that
watches every github project. We'll likely need to think a bit more
about how to enforce this on the github side as we do not want the
configuration to be managed by the project themselves.

Change-Id: I5ca6da61dc6ee250a2c6d451b09a0075fb1ff0b0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>