Update patch set 3
Patch Set 3:

So the global-jjb tox jobs does this annoying thing where it checks for changes in xml against expected-xml and forces the contributor to update expected xml files.

I'm starting to feel like not only is this inconvenient but makes it hard to contribute. I do like that it lets us know when patches are changing XML though.

I wonder if there's a better way we can report that information then failing a contributor's build.

Patch-set: 3
Label: Code-Review=0