Update patch set 5
Patch Set 5:

Now it looks good.

> it checks for changes in xml against expected-xml

I actually like that. It can still be improved.
Step 1: Have a script that re-generates the expected values.
Step 2: Hook that script to git-review somehow, so that it applies automatically before upload (not sure if that is compatible with Signed-off-by usage).

Keep the check to catch people which avoid the hook. The role of the auto-generated examples would be to show committers the real new xmls.

Patch-set: 5
Reviewer: Gerrit User 99 <99@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0