Add Maven job to build dependencies with patches
Provide a Maven job that can verify a patch including firstly building
projects it depends on with their patches first. This allows projects to
validate a patch that has upstream project dependencies which need to be
merged first.

Requires a new global variable to be defined in Jenkins GERRIT_URL
containing the base path to the Gerrit Web UI.

Issue: RELENG-526
Change-Id: Ie9ccb88c21ab5674d0156566da1daee4aa43d641
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>