Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

Ok I understand why verify jobs are failing the pattern now. It's because the merge job is has separate templates in it's name to handle multiple types but for some reason we decided to make the verify job a single job rather than multiple jobs so the verify is getting parsed poorly. If it's for efficiency because we don't want to launch many jobs then we do need a separate file-paths setting for the verify that's different then the merge job as done in this patch.

I still think we should be moving the file-paths into a {obj:var} setting as we do with Maven jobs that do this similarly. That would be an improvement. I am ok with merging this firstly and then working on the improvement as a separate patch though.

Patch-set: 2
Label: Code-Review=+2