Update patch set 6
Patch Set 6: Code-Review+1

Anil, updated the patch to include missing documentation for all of the defaults defined in the defaults section. Also made it possible to override the gerrit trigger file paths too.

If you're good with this we can merge this Monday and include it in the next global-jjb release.

Patch-set: 6