Convert packer file-paths cfg to {obj:var}
This is a standard we are using in most jobs so let's follow it with the
CI jobs as well. Looks like the packer jobs were already documented for
this but never actually implemented the feature.

The merge job unfortunately cannot be converted to the format due to
it's use of {platform} and {template} variables unless we want to make
it use the same file-paths as the verify job is now using.

Change-Id: Ie2976506e109f0e90ae26b4b6e4492d6b112b73a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>