Add note on JJB defaults
This patch provides additional clarification on how defaults are used in
JJB and when to expect the prioritization of variable expansion with
JJB.

Additionally add in intersphinx linking to jjb so that we can link
directly to JJB upstream documentation.

This documentation is based on details provided by Thanh Ha on the
OpenDaylight dev mailing list.

https://lists.opendaylight.org/pipermail/dev/2017-October/004184.html

Issue: RELENG-591
Change-Id: I65aaf933b88519dfbd1daaf8e00661146bc8c38f
Co-Authored-By: Anil Belur <abelur@linuxfoundation.org>
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>