Add install goal to javadoc job
It turns out javadoc needs to have the artifacts available to them in
order to run javadoc:aggregate. In some cases this means artifacts
installed to the local maven repo. This patch adds the install goal as
well as ensuring we continue to skip tests, checkstyle, and findbugs to
keep the jobs running quick.

Change-Id: I3ffa36b207a1412e0489d6c20a91ffcd3ac93c27
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>