Update patch set 7
Patch Set 7: Code-Review+1

> (1 comment)

Hello Thanh: The sample code should be comprehensive and self-sufficient to understand like we have in the JJB docs. If we refer to most of the examples are not complete such that they could be copied into a test.yaml to be tested on your CI. This is generic comment applicable to entire docs, not necessarily requires to be fixed in this change.

Patch-set: 7