Update patch set 7
Patch Set 7:

Hi Anil,

What would you recommend we change to make this more complete?

The docs here break down the 3 major sections, macros, templates, and projects (additionally defaults if used) which are all needed in order to make a working test. Together they produce something one can copy and paste into a test.yaml but requires the reader to read the docs in its entirety. Would creating a test.yaml source file that the reader can download with all 3 sections pre-filled after the Final thoughts be sufficient to satisfy this?

Keep in mind that this is a best practices documentation and is a complementary resource to the JJB docs and is not meant to be a complete replacement of official JJB documentation. It's intention is to describe the finer details of JJB usage as done in LF projects. Folks should still refer to the official JJB documentation to get the basic concepts.

Patch-set: 7
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0