Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

If a macro defines variables then when the macro is called the calling job-template MUST define all of the variables or redefine the variables.

This means if we want job to trigger, condition, and upstream jobs url to be configurable we have to define these variables again and set a blank default in every job template so that they are not required variables. Which means we probably want a non-generic variable to represent them so that it's clear that it's a configuration for the new publisher macro.

Patch-set: 3
Label: Code-Review=-1