Allow SONAR_MAVEN_GOAL to be configurable
Per reports on the mailing list Sonar jobs are broken in OpenDaylight.
This is because Sonar release a new plugin which only works with newer
versions of Sonar so until we can upgrade we need to allow flexibility
to configure the Sonar job to override the default sonar:sonar goal.

With this patch it is now possible to override the Sonar goal in JJB
(use defaults to globally override) and force an older version of the
Sonar plugin to run using a JJB defaults config or projects variable
like this:

  sonar-mvn-goal: org.codehaus.mojo:sonar-maven-plugin:3.3.0.603:sonar

Reports:
https://lists.opendaylight.org/pipermail/release/2017-December/013119.html
https://lists.opendaylight.org/pipermail/release/2017-December/013223.html

Change-Id: I04512613f39e7c328e0f2f18d9c4729667e86909
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>