Update patch set 4
Patch Set 4: Code-Review+1

Sadly to keep the job compatible with the existing deployment we are defining sonar-mvn-goal in the job-template. Since job template defaults have higher precedence in JJB than global-defaults it unfortunately means it is not possible to override the sonar-mvn-goal globally and must be done per-project. At least this gets us further than we were before where this was not possible to override at all.

Patch-set: 4
Label: Code-Review=+1