Add script to manage Jenkins cloud cfg
A groovy script to manage the Jenkins global cloud cfg
programmatically. This script reads configuration from
the directory jenkins-config/clouds to configure cloud
configuration in Jenkins. Refer to the documentation for
details.

No upgrade consideration is necessary as this script
automatically detects the absence of the jenkins-config/clouds
directory and disables the Cloud configuration.

Issue: RELENG-804
Change-Id: Idad3078942618b7d8fd08a40bf9e4e5ed5a0dc48
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>