Update patch set 2
Patch Set 2:

We should make sure we extensively test this before deploying. Although I think this patch is incomplete without providing a way for projects to choose their Java version in our maven-* jobs. We need to add to the scripts a call to "alternatives" to update and set the JAVA_HOME somewhere.

Patch-set: 2