Update patch set 5
Patch Set 5:

With this patch I get: /usr/lib/jvm/java-1.8.0-openjdk has not been configured as an alternative for java_sdk_openjdk

Which might not be a big deal considering we might not need java_sdk_openjdk and can likely be silently ignored. There's no failures for the other alternatives declarations.

Need to try with a java version that's not java-1.8 though....

Patch-set: 5