Update patch set 7
Patch Set 7:

> In my test, --install actually makes the right change..
 > 
 > Shall we use --install rather than --set?
 > 
 > Let me read about both to see if there are many differences. In
 > packer we seem to be using --set..

I'm pretty sure we need both.

--install let's alternatives know that the directory exists
--set tell's alternatives to set a directory as the default java

It works in my docker container but when I run it on Jenkins I've noticed despite both commands going  through "java -version" shows that the default java version did not switch.

I suspect it might have something to do with sudo as "root" setting the config and the "jenkins" user not getting the update. In a normal system I would logout and back in to make sure things take effect but we can't do that on Jenkins.

Patch-set: 7