Update patch set 7
Patch Set 7:

> Will every project with a ci-management/jjb area need to gain a
 > jenkins-config/ subdirectory for this to succeed?

Yes this will cause all projects to require a jenkins-config directory with a global-vars file defined for 'production' and 'sandbox' systems. (Unless they decide to exclude the job). This provides a very visible feature for the projects though so I don't really expect folks will want to do that and LF staff will need to fill out the initial details.

Patch-set: 7