Update patch set 1
Patch Set 1: Code-Review-1

Ah we also use this for logs. This is probably what's causing the failure for you. You have a system that does not have a Nexus backed logging server. We don't want to quietly ignore this though as we need to to throw a failure so that we know logs isn't working.

I think maybe the real solution here is to provide a way to deactivate logging for projects that don't need it.

Does this make sense?

Patch-set: 1
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1