Rename maven-release job to maven-stage
ATTENTION: This is a breaking change. The trigger word
"build release$" is changed to "stage-release$" to better
reflect what this job is actually doing.

In preparation for work on RELENG-425 this change renames the
maven-release jobs to maven-stage job to better reflect what the
job really does.

Trigger keyword was also changed to "stage-release$"

Issue: RELENG-425
Change-Id: I0dd4342664251386ddc7eca52df56156105f200d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>