Configurable Timed Trigger for Packer Merge Jobs
This change add the optional parameter 'cron' to packer builds so users
can configure when packer images can be rebuilt. It also switches the
timer from '00 10 1 * *' to '@monthly' to take advantage of Jenkins cron
hash system so the jobs are splayed over the month and don't all start
on the same minute.

Change-Id: I5bc29dab281d8ad1a4260994c1b1d6fb81b5a488
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>