Remove maven_metadata_validate function
This function was originally added to detect a bug with Maven 3.5.1
and earlier. At the time the OpenDaylight community was not yet ready
to migrate to 3.5.2 from 3.3.9 so we implemented this function as a
way to detect the metadata issues.

This function ended up causing some edge case issues and is not worth
maintaining. Remove the function and recommend all projects use Maven
3.5.2 as a minimum.

The specific issue in Maven was that sometimes the deploy phase would
generate an off by one timestamp mismatch between the metadata and the
filename of the artifacts. This causes invalid metadata and then
builds against affected artifacts fail.

Issue: RELENG-927
Change-Id: I65f4f6dc73382b91940620aba587f5fe56259eb1
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>