Create global tox.ini
Use -q to hide tox output.

Creates a global tox.ini which can be symlinked by projects to
automatically get commonly used tools in LF ciman repos. This replaces
the need to "source /path/to/venv" everywhere in our scripts to jump
to different tools. Instead we replace it with one of:

    tox -q -e openstack -- {parameters}
    tox -q -e openstack {parameters}

Non-CI management jobs can wget the tox.ini from global-jjb and reuse
it if they like.

Change-Id: I8b1a0936220ddbeb7f5e051ce15c421fec27c1a0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>