Setup tools in global-jjb tox.ini
Use -q to hide tox output.

Creates a global tox.ini which can be used by projects to use common
PyPi tools. This replaces the need to "source /path/to/venv"
everywhere in our scripts to jump to different tools.
Instead we replace it with one of:

    tox -q -c jjb/global-jjb/tox.ini -e openstack -- {parameters}
    tox -q -c jjb/global-jjb/tox.ini -e openstack {parameters}

Non-CI management jobs can wget the tox.ini from global-jjb to pull
the configuration.

By default the version installed will be the latest patch version in
semvar. The version comparison operator can be overrided via the
*_VERSION_OP environment variable for each tool type. Install versions
are also overridable via the *_VERSION environment variable.

Ex:

    JJB_VERSION=2.0.6
    JJB_VERSION_OP=">="

https://www.python.org/dev/peps/pep-0440/#version-specifiers

Change-Id: I8b1a0936220ddbeb7f5e051ce15c421fec27c1a0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>