Use --jobs-only option with jjb-deploy
JJB 2.0.3 implementation allows views to be managed by
code. The default behaviour is to update the job along with
all the views in the repository. This takes a longer if
the project has many views managed by code.

Therefore, use the flag --jobs-only to updated
on a given job in sandbox.

Change-Id: Ifaaf60f3479e7485e1dbb5618c6cbcce31525278
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>