Update patch set 1
Patch Set 1:

> > > Was there a reason for adding in this suffix handler? We had
 > > pretty
 > > > much left it out to force everyone to use the exact same yaml
 > > > suffix.
 > >
 > > Force everyone to use the suffix? Seems odd not to support
 > > something that's officially allowed and common practice.
 > 
 > Partly because it's the recommended extension for the files:
 > 
 > http://yaml.org/faq.html
 > 
 > I don't want to deal with our files ended up all named differently,
 > I like the idea of forcing folks to use the recommended extension.
 > 
 > Additionally, the upstream JJB project only allows in .yaml files
 > (look over their repo) I would prefer to not deviate from that.

I would rather not continue the debate 

> https://stackoverflow.com/a/38754994

that has gone on for years 

> http://markdblackwell.blogspot.com/2013/07/use-file-extension-yml-for-yaml.html

over whether .yaml or .yml is the preferred suffix.

I'm perfectly happy and fine with files within global-jjb ending in .yaml and us enforcing that. What I'm not okay with is adding more barriers for communities to use helpful tools, especially since there is quite a lot of work needed to use global-jjb in the first place, and it is such a trivial amount of work on our part.

Patch-set: 1
Reviewer: Gerrit User 15 <15@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0