Update patch set 1
Patch Set 1: Code-Review-1

> > How are you expecting to pass the random directory to later
 > steps?
 > 
 > I don't see '/tmp/v/jjb' used anywhere else, so I'm not sure which
 > later step would use this? From what I can see all the build steps
 > related to jjb happen after this on the same node: so once the venv
 > is created and activated, there's nothing left for them to do.

I just looked at how the jjb-install.sh is being used, you're correct.
I was misremembering how the verify and merge jobs go together.

So, the only thing outstanding is the tmpdir cleanup that needs to be
added as an extra script and step for those jobs.

Downgrading to -1 as it's not as breaking as I thought

Patch-set: 1
Label: Code-Review=-1