Update patch set 1
Patch Set 1:

(2 comments)

If we merge this, we should update all of ODL's jobs to support both recheck and reverify too before upgrading to whichever version of global-jjb this lands in otherwise if someone types "reverify" in ODL they will only run global-jjb jobs and again cause confusion on the ODL side.

(This was the reason we originally dropped reverify in ODL)

However if we can ensure all the regexes are the same everywhere I don't see this as a problem.

One suggestion to use '(recheck|reverify)$' as the search pattern, both reducing the number of lines added as well as makes it easier for folks who are copy pasting to get the full regex.

Patch-set: 1
Label: Verified=0