Update patch set 1
Patch Set 1:

Luis, what's the reasoning behind this? I see the mail thread that Thanh linked, but I don't understand the want to use the 'build-job' keyword on a project that isn't the ci management repo (releng/builder for ODL) since it's going to trigger a deployment of the jobs declared in the comment.

Patch-set: 1
Label: Verified=0