Consolidate all ssh keypair default names to jenkins-ssh
This is an attempt to consolidate all our different instances
of the ssh-keypair name all over the place. By default we will
consolidate on `jenkins-ssh` as the default keypair name. This
can be overrided of course and applies to the following config
locations in Jenkins / CIMAN:

1. defaults.yaml's `jenkins-ssh-credential`
2. Jenkins > Credentials > ID > jenkins-ssh
3. jenkins-config/clouds controlling both local privaate key name
   and public key name inside the cloud which now need to match

On a new infra if we are not overriding any values all of these should
accept the default value of `jenkins-ssh`.

Change-Id: I9fe691bbb59d81a1e861261be33cbda00de7a061
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>