Update patch set 11
Patch Set 11: Code-Review-1

(3 comments)

I've relicensed the files.

One thing I'd like to note about this change is that we may be imposing a defacto-standard on projects they did not agree to. I think it's fine to move this forward with the caveat that we start bringing in project in to comment on the format of the INFO.yaml file, and include some disclaimer that the format is still open for discussion.

Given this schema came out of OPNFV, and meets OPNFV needs, I think its reasonable to assume not all projects will/may have the same needs in what they need to track or list here.

Patch-set: 11
Label: Code-Review=-1