Update patch set 3
Patch Set 3:

Personally I think we should be setting the minimal accepted version of packer in our examples so that users of global-jjb can make a conscious decision of if they need to update or not rather than assuming everyone can upgrade to the latest version.

Keeping in mind that LF managed projects are no longer than only projects using global-jjb we need to be aware of our users.

Patch-set: 3