Update patch set 2
Patch Set 2:

(1 comment)

I'm still not convinced this will work as a bash alias, given the intricacies of shell and environment invocation.

Could the script just contain the call
> while [ $(fuser ..) ]; do sleep 1; done
and be executed?

Since 1) unattended upgrade should already be disabled on the images we run, and 2) any apt process running after this point should be from either the jenkins or build scripts.

If we continue seeing errors after that point, then we know something is wrong with how apt is configured in the image and it should be fixed in the Packer builds

Patch-set: 2
Reviewer: Gerrit User 15 <15@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0