Update patch set 2
Patch Set 2:

I was talking to Tony Espy from Canonical, apparently this issue is almost always the result of having "unattended-upgrades" installed.  Basically, there will be an unattended update going on in the background on boot.  He recommended I uninstall it on the packer Ubuntu snap image I made and the issue immediately disappeared.

Patch-set: 2