Update patch set 1
Patch Set 1: Code-Review-1

The CFG files are unused by the jjb-verify job thus would accomplish absolutely nothing in terms of verifying if the change is good or bad.

We need to develop some kind of linter to actually verify the integrity of the CFG files for a verify job to be of any value.

Patch-set: 1
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1