Move openstack-cron job to global-jjb
Move OpenDaylight's builder-openstack-cron job to global-jjb. This is
an initial patch that only migrates the "image-protection" script.

image-protection is used to parse jenkins-config/ and jjb/ directories
for any mention of images starting with "ZZCI" and flagging them as
protected. The assumption here is that ZZCI images are CI managed
and if they are referred to anywher by anything then they are
effectively considered in-use and should not be deletable so flag
them as "protected: true" in OpenStack.

Change-Id: Ib706663f0c08bd7a9160b0a6ce69dcd3b802f8f5
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>