Add Jenkins Sandbox Cleanup job
This cleanup job runs weekly on Saturdays to clear out Sandbox
jobs and views from the system.

The benefit of this is that it uses the same version of JJB as
the other CI jobs whereas the traditional cron job we put in puppet
does not often get updated to match the right version.

Issue: RELENG-1118
Change-Id: I2229d66f637bd5aab203a3bbf12aa1e677a471a1
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>