Update patch set 1
Patch Set 1:

I'm really on the fence with this change. It only solves 1/2 of the configuration problem. That of the Jenkins jobs. It does nothing for the Nexus side for clearing out the job logs that exist in the sandbox holding area which continues to have to be a cron job.

Patch-set: 1
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Verified=0