Update patch set 1
Patch Set 1:

> I'm really on the fence with this change. It only solves 1/2 of the
 > configuration problem. That of the Jenkins jobs. It does nothing
 > for the Nexus side for clearing out the job logs that exist in the
 > sandbox holding area which continues to have to be a cron job.

To be fair the Nexus side is a completely separate step. Even if we ignore this job. It required modifying 2 separate puppet files. One for Jenkins and one for Nexus. Irrespective of this patch it was still a 2 step process previously too. This reduces the steps to 1 but does bring up a point that we need to document the Nexus part in our infrastructure docs.

Patch-set: 1
Label: Code-Review=0